multi: make asset optional in proof file
Because in case of an interactive full value send the sender will get a
proof that shows they don't have any asset anymore. But in that case the
Asset field in the proof struct cannot be set and therefore needs to be
a pointer so we can set it to nil.