proof: add test for verifying proof file
This test is added mainly as an easy way to decode and verify a proof
file encoded as a hex string.
This allows us to step through the validation process of a proof file in
a debugger if we ever encounter an invalid proof file during development
(for example when writing a new itest).