multi: fix manual grouped asset creation in tests
In this commit, we update multiple unit tests to not set an asset group
key directly, and instead use some test helper that calls asset.New().
This ensures that the group witness is copied into the asset. The only
time setting the group key directly is acceptable is before a group
witness has been created, or in a minting request.